% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBathym.R
\name{getBathym}
\alias{getBathym}
\title{Extract bathymetry}
\usage{
getBathym(points, desired.diameter = 0.7, func = "mean")
}
\arguments{
\item{points}{Data frame. Input data frame with columns labeled \code{lon} and \code{lat}}

\item{desired.diameter}{Numeric. Defaults to 0.7˚. Results in a 7x7 (= 49) pixel box (0.7x0.7˚) around the central location.}

\item{func}{Character. Default "mean", can change to "sd".}
}
\value{
A five column data frame with bathymetry extracted for lat/lons of interest.
}
\description{
This function extracts bathymetry for lat/lons of interest directly from ERDAPP.
}
\details{
Can be slow for large datasets.
}
\examples{
points <- expand.grid(
  lon = seq(-132, -122, by = 2),
  lat = seq(32, 40, by = 2)
)

testbathym <- getBathym(
  points = points,
  desired.diameter = 0.7,
  func = "mean"
)

head(testbathym)
}
